token_api/requirements.txt
fastapi
uvicorn
web3
pyjwt
python-dotenv
